More Terrain Work
Added in smoother, rolling hills in the dry area. Added some interesting ridge-like hill features. Added in the two upper waterfalls and the spray effects for when they hit the water. Added in spray effects for the lower waterfalls flowing over. Adjusted the water level and speed effects again. Did a small bit of smoothing the some jagged areas.