More Terrain Work
Added in the big mountain to the skybox, but wasn't able to add in the cloudy sky yet. Added a bit more trees and grass and adjusted their settings so that the grass renders in at farther distances and the trees cull out at far distances in order to reduce computation consumption and improve framrate and playability. Recolored the waterfalls to get their color a bit closer to that of the water. Smoothed out some more jagged terrain areas.