More POI Additions
UV wrapped hay texture to the Big Nest. Added texture to the Tree Stump. Added in the eggs to the Nesting Grounds. Added colliders to the eggs, Big Nest, Tree Stump, Ancient Tree, Hollow Log, and Floating Log. Made the Floating Log bob up and down in the water with a custom script.